Made the login page fully functional
[+] Created AuthContext to manage and persist user data in localStorage
[+] Implemented the API client
[+] Created new ErrorBlock component to display error messages on the page
[+] Implemented useAuth hook to handle auth operations with the API
[o] Modified i18n language files related to auth
[o] Changed the BIGINT based IDs to INTEGER in the database so that pg can properly parse numbers
[o] Changed the login endpoint to return user data instead of a 204 No Content response
[o] Added the AuthProvider and Toaster inside of App
[o] Made the login page functional
[o] Fixed GenericButton not triggering the click event

+++ PostgreSQL/10_users.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     first_name TEXT NOT NULL,
     last_name TEXT NOT NULL,
     email TEXT NOT NULL,
@@ -10,7 +10,7 @@ CREATE TABLE users (
 
 CREATE TABLE sessions (
     session_id VARCHAR(64) PRIMARY KEY,
-    user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     expires_at TIMESTAMP NOT NULL DEFAULT (NOW() + INTERVAL '30 days')
 );
